Added proto file, grpcService for UserManagement, Modified previous services and added Auto Scheduler for Deadline Orders

+++ Logistics/Services/OrderService.cs
@@ -26,30 +26,36 @@ namespace Logistics.Services
         public async Task AddNewAvailableOrderAsync(Available order)
         {
             await _available.InsertOneAsync(order);
+            await _deadline.InsertOneAsync(new Deadline { Id = order.UniqueId, DateTimeFrom = order.DateTimeFrom });
         }
 
         public async Task<List<Deadline>> CheckDeadlineOrdersAsync()
         {
             return await _deadline
-                .Find(order => order.DateTimeTo <= DateTime.UtcNow)
+                .Find(order => order.DateTimeFrom <= DateTime.UtcNow)
                 .ToListAsync();
         }
 
         public async Task DeleteOrderByIdAsync(Guid id)
         {
+            if (await GetOrderByIdAsync(id) is null)
+            {
+                throw new ArgumentException("No Order with such Id");
+            }
+            await _deadline.DeleteOneAsync(d => d.Id == id);
             await _available.DeleteOneAsync(order => order.UniqueId == id);
         }
 
         public async Task<List<Available>> GetAllOrdersAsync(OrderQuery query)
-        {            
+        {
             var orders = _available.AsQueryable();
 
-            if (!string.IsNullOrWhiteSpace(query.AddressFrom)) 
+            if (!string.IsNullOrWhiteSpace(query.AddressFrom))
                 orders = orders.Where(o => o.AddressFrom.Contains(query.AddressFrom));
-            
-            if (!string.IsNullOrWhiteSpace(query.AddressTo))            
+
+            if (!string.IsNullOrWhiteSpace(query.AddressTo))
                 orders = orders.Where(o => o.AddressTo.Contains(query.AddressTo));
-            
+
             if (query.DateTimeFrom.HasValue)
                 orders = orders.Where(o => o.DateTimeFrom >= query.DateTimeFrom.Value);
 
@@ -59,39 +65,39 @@ namespace Logistics.Services
             if (!string.IsNullOrWhiteSpace(query.FrachtType))
                 orders = orders.Where(o => o.FrachtType == query.FrachtType);
 
-            if (query.Distance.HasValue)            
+            if (query.Distance.HasValue)
                 orders = orders.Where(o => o.Distance == query.Distance.Value);
-            
-            if (!string.IsNullOrWhiteSpace(query.TrunkType))            
+
+            if (!string.IsNullOrWhiteSpace(query.TrunkType))
                 orders = orders.Where(o => o.TrunkType == query.TrunkType);
-            
+
             if (query.Weight.HasValue)
                 orders = orders.Where(o => o.Weight == query.Weight.Value);
-            
+
             if (query.LoadingMetre.HasValue)
                 orders = orders.Where(o => o.LoadingMetre == query.LoadingMetre.Value);
-            
+
             if (query.Height.HasValue)
                 orders = orders.Where(o => o.Height == query.Height.Value);
-            
+
             if (!string.IsNullOrWhiteSpace(query.LoadingType))
                 orders = orders.Where(o => o.LoadingType == query.LoadingType);
-            
+
             if (query.Temperature.HasValue)
                 orders = orders.Where(o => o.Temperature == query.Temperature.Value);
-            
+
             if (query.Price.HasValue)
                 orders = orders.Where(o => o.Price == query.Price.Value);
-            
+
             if (!string.IsNullOrWhiteSpace(query.ContactInfo))
                 orders = orders.Where(o => o.ContactInfo.Contains(query.ContactInfo));
-            
+
             return await orders.ToListAsync();
         }
 
-        public Task<Available> GetOrderByIdAsync(Guid id)
+        public async Task<Available?> GetOrderByIdAsync(Guid id)
         {
-            return _available.Find(order => order.UniqueId == id).FirstOrDefaultAsync();
-        }       
+            return await _available.Find(order => order.UniqueId == id).FirstOrDefaultAsync();
+        }
     }
 }
